fix(clarity): dual React instance SSR crash and node:child_process in client bundle
Fixes #124

Two issues preventing clarity UI from loading:

1. sonner resolved to root react (v19.2.0) while SSR used local react-dom
   (v19.2.4) — added sonner to clarity UI's own package.json so both
   resolve the same instance.

2. FillWeekCard imported getDaysInPeriod/subtractDay from src/utils/date.ts
   which had a top-level `import { execSync }` — Vite externalized
   node:child_process for the browser, crashing at runtime. Extracted
   locale detection into date-locale.ts and switched to lazy require()
   so date.ts has zero static Node.js imports.
