+++ src/clarity/ui/package.json
@@ -21,6 +21,7 @@
 		"cmdk": "^1.1.1",
 		"lucide-react": "^0.544.0",
 		"react": "^19.2.1",
+		"sonner": "^2.0.7",
 		"react-dom": "^19.2.1",
 		"tailwind-merge": "^3.0.2",
 		"tailwindcss": "^4.0.6"
